refactor: break out some logic into helpers module, whitespace fixes
