+++ app/__init__.py
@@ -1,2 +1,2 @@
-from .menu import Menu
 from .app import PACKAGE_DIR, InsideTheBoxTUI
+from .menu import Menu
