--- app/menu/__init__.py
@@ -1,1 +1,0 @@
-from .menu import Menu
